LDEV-3776: Clean up and format all LAMS Java code.

+++ lams_monitoring/src/java/org/lamsfoundation/lams/monitoring/dto/LearnerProgressBatchDTO.java
@@ -44,8 +44,8 @@ public class LearnerProgressBatchDTO {
 	this.numAllLearnerProgress = numAllLearnerProgress;
 
 	// integer division truncates, so convert to floating point first.
-	double batchSizeAsDouble = (double) batchSize;
-	double numAllLearnerProgressAsDouble = (double) numAllLearnerProgress;
+	double batchSizeAsDouble = batchSize;
+	double numAllLearnerProgressAsDouble = numAllLearnerProgress;
 	double numBatchesDouble = numAllLearnerProgressAsDouble / batchSizeAsDouble;
 	numBatchesDouble = Math.ceil(numBatchesDouble);
